// ui/lib/api/schemas/llmProviderCatalog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 */
import type { LlmCatalogModel } from './llmCatalogModel'
import type { LlmProviderCatalogStatus } from './llmProviderCatalogStatus'

export interface LlmProviderCatalog {
  /** @nullable */
  baseUrl?: string | null
  /** @nullable */
  error?: string | null
  hasApiKey: boolean
  id: string
  models: LlmCatalogModel[]
  name: string
  requiresApiKey: boolean
  requiresBaseUrl: boolean
  status: LlmProviderCatalogStatus
}